- Changed: GameObject.Realm has now changed from a byte to an eRealm (which can still be cast to a byte).
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@954 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/behaviour/Requirements/RealmRequirement.cs
@@ -73,11 +73,9 @@ namespace DOL.GS.Behaviour.Requirements
 			bool result = true;
             GamePlayer player = BehaviourUtils.GuessGamePlayerFromNotify(e, sender, args);
 
-            result = compare(player.Realm, N, Comparator);
+            result = compare((int)player.Realm, (int)N, Comparator);
 
 			return result;
 		}
-
-		
     }
 }
