+++ GameServer/quests/Quests/QuestScripts/albion/RevengeTheOtherWhiteMeat.cs
@@ -145,7 +145,7 @@ namespace DOL.GS.Quests.Albion
 				if (log.IsWarnEnabled)
 					log.Warn("Could not find " + farmerAsma.Name + ", creating him ...");
 				farmerAsma.GuildName = "Part of " + questTitle + " Quest";
-				farmerAsma.Realm = (byte) eRealm.Albion;
+				farmerAsma.Realm = eRealm.Albion;
 				farmerAsma.CurrentRegionID = 1;
 
 				GameNpcInventoryTemplate template = new GameNpcInventoryTemplate();
@@ -485,7 +485,7 @@ namespace DOL.GS.Quests.Albion
 						GameNPC pigHerderWyatt = new GameNPC();
 						pigHerderWyatt.Model = 39;
 						pigHerderWyatt.Name = "Pig Herder Wyatt";
-						pigHerderWyatt.Realm = (byte) eRealm.Albion;
+						pigHerderWyatt.Realm = eRealm.Albion;
 						pigHerderWyatt.CurrentRegionID = 1;
 
 						GameNpcInventoryTemplate template = new GameNpcInventoryTemplate();
